Update README.md
The Text Summarizer's goal is to build a chatbot that can take user input in the form of text and return a summarized version of that text using NLTK (Natural Language Toolkit). The chatbot will leverage natural language processing (NLP) techniques to identify the most important sentences in the text and generate a concise summary.